DangerFloor ready for sprites and level design

+++ Assets/Scripts/DangerFloor.cs
@@ -3,7 +3,7 @@ using System.Collections;
 
 public class DangerFloor : MonoBehaviour {
 
-	private float DAMAGE;
+	private int DAMAGE;
 
 	// Use this for initialization
 	void Start () {
